Integrate OnboardingScreen into MainActivity
The `MainActivity` now displays an `OnBoardingScreen` if the user has not seen it before.
The start destination of the `NavHost` is determined by the `isOnboardingShown` state, which is collected from `AppPreferencesService`.
After completing the onboarding process, the app navigates to the "home" screen and removes "onboarding" from the back stack.
The `HomeScreen` composable has also been added to the "home" route.
An `@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")` has been added to suppress a lint warning.