Feat: Integrate Onboarding Screen (#59)
* Inject TaskRepository into OnboardingViewModel

The `OnboardingViewModel` now receives `TaskRepository` as a dependency via constructor injection. This change is reflected in the Koin `viewModelModule`.

* Remove status and navigation bar padding from OnBoardingScreen

The `OnBoardingScreen` now uses `koinViewModel()` to obtain the `OnboardingViewModel` by default.

* Integrate OnboardingScreen into MainActivity

The `MainActivity` now displays an `OnBoardingScreen` if the user has not seen it before.
The start destination of the `NavHost` is determined by the `isOnboardingShown` state, which is collected from `AppPreferencesService`.
After completing the onboarding process, the app navigates to the "home" screen and removes "onboarding" from the back stack.
The `HomeScreen` composable has also been added to the "home" route.
An `@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")` has been added to suppress a lint warning.

* Refactor: Defer pre-defined category creation until onboarding is confirmed

This commit modifies the `MainActivity` to defer the creation of pre-defined categories until it's confirmed that the onboarding screen needs to be shown.

Previously, `createPreDefineCategories()` was called unconditionally in `onCreate()`. Now, it's only invoked when the `isOnboardingShownState` indicates that onboarding has not been completed.

Additionally, a loading indicator (`CircularProgressIndicator`) is displayed while the `isOnboardingShownState` is being determined to provide visual feedback to the user. The "category" route now navigates to `CategoriesScreen()`.

---------

Co-authored-by: Khaled.Eid <khaled.eid1k1k@gmail.com>