Refactor: Defer pre-defined category creation until onboarding is confirmed
This commit modifies the `MainActivity` to defer the creation of pre-defined categories until it's confirmed that the onboarding screen needs to be shown.

Previously, `createPreDefineCategories()` was called unconditionally in `onCreate()`. Now, it's only invoked when the `isOnboardingShownState` indicates that onboarding has not been completed.

Additionally, a loading indicator (`CircularProgressIndicator`) is displayed while the `isOnboardingShownState` is being determined to provide visual feedback to the user. The "category" route now navigates to `CategoriesScreen()`.